requirements.txt
streamlit
pandas
scikit-learn
langchain
sentence-transformers
faiss-cpu
